backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.9
numpy==1.24.2
opencv-python==4.7.0.72
matplotlib==3.7.1
scikit-learn==1.2.2
pandas==2.0.0
scipy==1.10.1
pillow==9.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
